requirements.txt
aiogram
paho-mqtt>=2.1.0
dotenv
asyncio
pyyaml